% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{count_donations}
\alias{count_donations}
\title{Group donations by sector and return summary}
\usage{
count_donations(donations = get_donation_data())
}
\arguments{
\item{donations}{a data frame created by \code{\link[=get_donation_data]{get_donation_data()}}}
}
\value{
a tibble or data.table
}
\description{
Group donations by sector and return summary
}
\examples{

count_donations()

}
